requirements.txt
Django==4.0.2
django_cors_headers==3.11.0
djangorestframework==3.13.1
drf_nested_routers==0.93.4
djangorestframework-xml==2.0.0